Cleanup code: move private fields to top, create file for LogStore class
