--- ios/Flutter/Generated.xcconfig
@@ -1,15 +1,0 @@
-// This is a generated file; do not edit or check into version control.
-FLUTTER_ROOT=/Users/zebul/flutter
-FLUTTER_APPLICATION_PATH=/Users/zebul/Projects/logbrowser
-FLUTTER_TARGET=/Users/zebul/Projects/logbrowser/lib/main.dart
-FLUTTER_BUILD_DIR=build
-SYMROOT=${SOURCE_ROOT}/../build/ios
-OTHER_LDFLAGS=$(inherited) -framework Flutter
-FLUTTER_FRAMEWORK_DIR=/Users/zebul/flutter/bin/cache/artifacts/engine/ios
-FLUTTER_BUILD_NAME=1.0.0
-FLUTTER_BUILD_NUMBER=1
-DART_DEFINES=flutter.inspector.structuredErrors%3Dtrue
-DART_OBFUSCATION=false
-TRACK_WIDGET_CREATION=true
-TREE_SHAKE_ICONS=false
-PACKAGE_CONFIG=/Users/zebul/Projects/logbrowser/.dart_tool/package_config.json
